requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
selenium==4.15.2
webdriver-manager==4.0.1
numpy==1.26.0
pandas==2.1.1
openpyxl==3.1.2
python-dotenv==1.0.0
names==0.3.0
Flask-SocketIO==5.3.6
eventlet==0.33.3
requests==2.31.0
